logproto-client-factor: introduce a stateful flag
This new stateful flag is going to describe if the LogProtoClientFactory is going to produce a stateless or stateful LogProtoClient.
Normally a LogProtoClient is stateless (prefered), but it is possible to create a LogProtoClient that track the sent messages.
Later the driver could make decisions based on this LogProtoClient attribute.

Signed-off-by: Kokan <kokaipeter@gmail.com>