logwriter: rewind backlog when proto disconnected
In case of a partial write the LogProto stores it in an internal state (partial) this fact.
The partial message is neither acked nor rewinded, only when it was sent completly.
If the connection is closed, when the message is partialy sent the LogProto is freed, thus a partial state of the message as well.
As the result a message is neither acked nor rewinded ever.

This could be easily observed with reliable disk queue, where a message could stuck in the backlog.

Signed-off-by: Kokan <kokaipeter@gmail.com>