filter: filter_parser outputs LogFilterPipe
The filter_parser output was FilterExprNode, which later was encapsulated with LogFilterPipe.
The aim of this whole patchset to make FilterExprNode extinct and move the structe it holds into a LogExprNode structure.
(merge the two graph into one)

This patch does not achieve this, as only single filter (without or, and, (), not) are accepted.
Also as it turned out *filter*s are used in random places, where the LogFilterPipe is not really needed. This *probably* could kill this solution, it needs some rethinking!!!

Signed-off-by: Kokan <kokaipeter@gmail.com>