syslogformat: request handles at every module init
The syslogformat module caches handles at module initialization time. Caching
those handles are possible due to they are global, and do not change while
syslog-ng is running, even in case of reload those handle values are still not
change. As those values are stored and assigned by the logmsg_registry.
The logmsg_registry assignes handle in the order they first apear, therefore
those value might differ in case of restart. The logmsg_registry is initialized
in app_startup, and destroyed in the app_shutdown.

Based on that caching must be fine as long as app_shutdown is not called, that is
not called in the main application only just before stopping the execution of the
process. But this could have been an issue in case of unit tests. Fortunatly the
current Criterion implementation is fine (Note2).

In case of syslogformat init there is also a flag, so these messages are not
requested multiple time. As there is no module deinit, the flag once turned on,
always keeps that propery. The module init is a rare event compared to message
process, therefore doing multiple query towards logmsg_registry is fine, and
solves the issue of caching invalid value.

Note1: Serialized message format such as the one used in diskq, must update the
       handles when the process starts or read those data the first time.
Note2: The question would be valid why this does not cause any issue with unit
       tests ? This is due to Criterion forks for each test execution as of now.

Signed-off-by: Kokan <kokaipeter@gmail.com>