nvtable: fix off-by-one in checking whether a handle is static or dynamic
This patch fixes two off-by-one cases in nvtable by migrating to the use
of the newly extract nv_table_is_handle_static() function.

As static/dynamic handles are stored/serialized differently this
patch merits a deeper look if we break anything.

Background
==========

the nvtable code differentiates between "static" and "dynamic" handles
based on the value of "num_static_entries". The major difference between
the two is that:
  - static entries don't store their "name" in the NVTable
    (assuming that the handle alone will be a static/unique identifier
    of the name-value pair)
  - static entries are using a simple array as a look up table, indexed
    by the handle (in contrast to a dynamic entry, which uses a sorted
    array of (handle, offset) pairs), on which we performing binary search

The bug
=======

We had various points in the code that checked if a handle was "static"
or "dynamic". Some of these conditions considered the handle that equals to
num_static_entries as a "static" handle, others considered it
as a "dynamic" handle.

This is being fixed by extracting a decision into a
nv_table_is_handle_static() function.

Both of the actual bugs (in nv_table_add_value() and nv_table_set_indirect_entry())
considered the last handle as "dynamic" instead of "static" as it
should have. This means that they stored the "name" for the handle in the
entry.

Potential consequences of the fixing of the bug
===============================================

Since dynamic handles do have "name" in the NVTable representation exactly
where we store the "value" in "static" cases, this can become
problematic, especially as we store the same representation on disk when
we serialize a LogMessage.

Why it isn't a problem in this case
===================================

There are multiple reasons:

Because we are not using handle "9" ever:
* num_static_entries is initialized to LM_V_MAX, which is 9 in the current
  versions of syslog-ng.
* the handle "9" is associated to the very first macro (by chance), which is "FACILITY"
* macros are never stored in NVTable entries

If we were using it (despite the explanation above):

Direct entries:
* Due to the bug in nv_table_add_value() would add the "name" to the variable part, but at the same
  time it would set entry->name_len, basically storing the "name" in the payload
  against our original desire, increasing the NVTable size.
* but whenever we resolve that value, we'd just ignore the name (due to nv_table_resolve_entry() properly handling the case)

Indirect entries:
* Due to the bug in nv_table_set_indirect_entry(), we would add "name" to the
  variable portion of indirect entries, which again would be handled properly.

Also (due to another unrelated bug), the allocation of the NVEntry for
handle 9 will allocate enough for the name being stored (even though this
is exactly what we wanted to avoid).

All in all, even though this was an off-by-one issue, this wasn't actually
triggered and even if it were we are ok.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>