nvtable: don't add the "name" value for static entries in case they are overwritten
This patch enables the reduction of the NVTable memory consumption (thus
LogMessage), because we erroneously allocated space for the "name" of
static (builtin) values and we also stored it there in case we were changing
an existing entry.

This part of the patch will removes the storing of the name, the followup
patch would decrease the memory allocated.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>