nvtable: improve performance of nv_table_add_value() when entries are added in sorted order
When the name-value pairs are added to a message in handle-sorted order, we
can have a short-cut instead of doing a full blown binary search: check if
the last element of the sorted array is smaller than the handle that need
finding.  If it is smaller, we can be sure that the new element is to be
placed last.

This improves performance by about 30% for the compaction performance tests.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>