block: support unnamed value("unnmaed" key(value))
The current block argument list must be in the key(value) format.
Example:
```
block root block-name(key(value), ...) { }
```

There are plugins that has values without a name, such as the *udp*, *tcp* source.
```
udp( "127.0.0.1" port(1234) );
```

In an attempt to rewrite such plugins based on other plugin is impossible as of now.
This change attempts to introduce a value only parameter for the block.

The current implementation does not require any extra modification, every block now has a new key entry called `__single_value`. The block writer could refer to it as usually: `__single_value`, and it is replaced with empty text or the content of the first value in the block such as:

Implementation:
```
block root value-only()
{
  `__single_value`
};
```

Usage:
```
value-only( "foo" );
```

Preprocessed into:
```
 "foo"
```

Note: the side effect is that now every block based plugin going to accept the first string as a single value.

Signed-off-by: Kokan <kokaipeter@gmail.com>