log_msg_append_format_sdata: enhance sd-id sd-name separation in case of enterpriseId
There are two formats for SD-ID names:
* Names that do not contain an at-sign ("@", ABNF %d64) are reserved
* Anyone can define additional SD-IDs using names in the format
  name@<private enterprise number>, e.g., "ourSDID@32473".

SDATA is stored as nvpair as string, and no structural information
saved. The way it currently works is looking up the last dot(.), and
claim everything before that as SD-ID. Resulting in invalid sdata, when
the PARAM-NAME has dot(.) in it, which is valid.

The following configuration results in invalidate sdata on the
destination:
```
@version: 3.30

log {
 source      { example-msg-generator(num(1)); };
 rewrite     { set("112" value(".SDATA.test@1.2.3.log.fac")); };
 destination { syslog("127.0.0.1" port(1111)); };
};
```

The fix depends on the fact that SD-ID has either contains at(@) or not,
when it contains at(@), it must followed by enterpriseId(number)
optionally with sub-id (number also).
As such the first non dot(.) or non digit shall be part of PARAM-NAME.

Signed-off-by: Kokan <kokaipeter@gmail.com>