light/python: add option mapping clazz -> class
There is a keyword collision in case of python parser, both python and syslog-ng has class as a keyword. In syslog-ng parser { python(class("PythonCodeClass")); };. In python the class also used, so the name cannot be reused as a function argument name. A mapping is done, clazz is rewritten to class.

Signed-off-by: Kokan <kokaipeter@gmail.com>